requirements.txt
yfinance
pandas
numpy
requests
beautifulsoup4
streamlit
plotly
supabase
python-telegram-bot
apscheduler
schedule
scikit-learn
xgboost
ta
transformers
torch
python-dotenv
feedparser